Create README.md
Welcome to Plane Warrior, the thrilling 2D aerial combat game that will take you on an adrenaline-fueled journey through the skies! Strap yourself in and prepare for an epic adventure like no other.

In Plane Warrior, you are a skilled pilot, tasked with defending your homeland from waves of enemy aircraft. As an ace flyer, you possess unmatched aerial maneuverability and a keen eye for precision targeting. It's up to you to become the ultimate hero, taking down enemy forces and restoring peace to your land.